refactor: move version resolver to common
Move the version resolver to `common` directory so that it can be shared
with client code.
