+++ client/tsconfig.json
@@ -15,7 +15,7 @@
     }
   ],
   "include": [
-    "src"
+    "src/*.ts"
   ],
   "exclude": [
     "node_modules"
