+++ jasmine.json
@@ -2,6 +2,7 @@
   "spec_dir": "dist",
   "spec_files": [
     "client/tests/*_spec.js",
+    "common/tests/*_spec.js",
     "server/tests/*_spec.js"
   ]
 }
