+++ server/banner.tsconfig.json
@@ -1,13 +1,16 @@
 {
+  "extends": "../tsconfig.json",
   "compilerOptions": {
-    "target": "es2015",
-    "module": "es2015",
-    "moduleResolution": "node",
-    "sourceMap": false,
-    "strict": true,
-    "outDir": "../dist/banner"
+    "outDir": "../dist/banner",
+    "rootDir": "src",
+    "rootDirs": [".", "../dist"]
   },
+  "references": [
+    {"path": "../common"}
+  ],
   "files": [
-    "src/banner.ts"
+    "src/banner.ts",
+    "src/cmdline_utils.ts",
+    "src/version_provider.ts"
   ]
 }
