+++ test.tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "./tsconfig.json",
+  "references": [
+    {"path": "./client/src/tests"},
+    {"path": "./server/src/tests"},
+    {"path": "./common/tests"}
+  ]
+}
